Avoid the warning 'QTableWidget: cannot insert an item that is already owned by another QTableWidget' that happens in Qt 4.4.2

+++ src/mytablewidget.cpp
@@ -19,6 +19,8 @@
 #include "mytablewidget.h"
 #include <QTableWidgetItem>
 
+#define BE_VERBOSE 0
+
 MyTableWidget::MyTableWidget( QWidget * parent ) : QTableWidget(parent) 
 {
 }
@@ -28,31 +30,49 @@ MyTableWidget::MyTableWidget( int rows, int columns, QWidget * parent )
 {
 }
 
-QTableWidgetItem * MyTableWidget::getItem(int row, int column ) {
+QTableWidgetItem * MyTableWidget::getItem(int row, int column, bool * existed ) {
+#if BE_VERBOSE
+	qDebug("MyTableWidget::getItem: %d, %d", row, column);
+#endif
 	QTableWidgetItem * i = item(row, column);
+	if (existed != 0) *existed = (i!=0); // Returns if the item already existed or not
 	if (i != 0) return i; else return createItem(column);
 }
 
 QTableWidgetItem * MyTableWidget::createItem(int /*col*/) {
+#if BE_VERBOSE
+	qDebug("MyTableWidget::createItem");
+#endif
 	QTableWidgetItem * i = new QTableWidgetItem();
 	i->setFlags(Qt::ItemIsSelectable | Qt::ItemIsEnabled | Qt::ItemIsDragEnabled);
 	return i;
 }
 
 void MyTableWidget::setText(int row, int column, const QString & text ) {
-	QTableWidgetItem * i = getItem(row, column);
+#if BE_VERBOSE
+	qDebug("MyTableWidget::setText: %d, %d", row, column);
+#endif
+	bool existed;
+	QTableWidgetItem * i = getItem(row, column, &existed);
 	i->setText(text);
-	setItem(row, column, i);
+	if (!existed) setItem(row, column, i);
 }
 
 QString MyTableWidget::text(int row, int column) {
+#if BE_VERBOSE
+	qDebug("MyTableWidget::text: %d, %d", row, column);
+#endif
 	return getItem(row, column)->text();
 }
 
 void MyTableWidget::setIcon(int row, int column, const QIcon & icon ) {
-	QTableWidgetItem * i = getItem(row, column);
+#if BE_VERBOSE
+	qDebug("MyTableWidget::setIcon %d, %d", row, column);
+#endif
+	bool existed;
+	QTableWidgetItem * i = getItem(row, column, &existed);
 	i->setIcon(icon);
-	setItem(row, column, i);
+	if (!existed) setItem(row, column, i);
 }
 
 QIcon MyTableWidget::icon(int row, int column) {
